requirements.txt
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
mdutils==1.3.1
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
PyExecJS==1.5.1
pyjsparser==2.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
